Introduce symposium-sdk crate and refactor main crate to use it
Add a new `symposium-sdk` crate as the single SDK dependency for plugin
binaries (hooks, custom predicates, subcommands). It re-exports
`symposium-hook` as `hook` and adds:

- `predicate` module: `PredicateOutput` and `SelectedCrate` types for
  custom predicate witness output serialization
- `workspace` module: `WorkspaceCrate` struct for typed workspace
  dependency access

Refactor the main symposium crate to use these SDK types as the source
of truth:
- `crate_sources::WorkspaceCrate` is now a re-export from the SDK
- `predicate::WitnessCrate` is now an alias for `SelectedCrate`
- `parse_witness_stdout` deserializes via `PredicateOutput` directly
