+++ Cargo.toml
@@ -44,6 +44,7 @@ sha2 = "0.10"
 tar = "0.4"
 tempfile = "3.6"
 symposium-hook = { path = "symposium-hook", features = ["clap"] }
+symposium-sdk = { path = "symposium-sdk" }
 tokio = { version = "1", features = ["macros", "rt-multi-thread"] }
 toml = "0.8"
 tracing = "0.1"
@@ -62,4 +63,4 @@ indoc = "2.0.7"
 symposium-testlib = { path = "symposium-testlib" }
 
 [workspace]
-members = [".", "symposium-testlib", "symposium-hook", "symposium-install"]
+members = [".", "symposium-testlib", "symposium-hook", "symposium-install", "symposium-sdk"]
