+++ src/predicate.rs
@@ -196,7 +196,7 @@ impl Predicate {
                 let witness = ctx.custom_witness(name, arg)?;
                 let pairs = witness
                     .iter()
-                    .map(|wc| (wc.name.clone(), wc.version.clone()))
+                    .map(|wc| (wc.crate_name.clone(), wc.version.clone()))
                     .collect();
                 Some(pairs)
             }
@@ -788,30 +788,10 @@ fn join(preds: &[Predicate]) -> String {
 // --- custom predicate evaluation infrastructure ---
 
 /// A crate reported by a custom predicate's witness output.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct WitnessCrate {
-    pub name: String,
-    pub version: semver::Version,
-}
-
-impl<'de> serde::Deserialize<'de> for WitnessCrate {
-    fn deserialize<D: serde::Deserializer<'de>>(
-        deserializer: D,
-    ) -> std::result::Result<Self, D::Error> {
-        #[derive(serde::Deserialize)]
-        struct Raw {
-            #[serde(rename = "crate")]
-            crate_name: String,
-            version: String,
-        }
-        let raw = Raw::deserialize(deserializer)?;
-        let version = semver::Version::parse(&raw.version).map_err(serde::de::Error::custom)?;
-        Ok(WitnessCrate {
-            name: raw.crate_name,
-            version,
-        })
-    }
-}
+///
+/// This is a re-export of [`symposium_sdk::predicate::SelectedCrate`] — the
+/// same type plugin authors serialize on stdout.
+pub use symposium_sdk::predicate::SelectedCrate as WitnessCrate;
 
 /// Cached result of a custom predicate invocation.
 #[derive(Debug, Clone)]
@@ -900,17 +880,13 @@ fn run_custom_predicate(
 /// predicate should be treated as failed). Returns `Some(vec![])` for empty
 /// stdout (pass, no witness crates). Returns `Some(crates)` on valid JSON.
 fn parse_witness_stdout(predicate_name: &str, stdout: &[u8]) -> Option<Vec<WitnessCrate>> {
+    use symposium_sdk::predicate::PredicateOutput;
+
     if stdout.is_empty() {
         return Some(Vec::new());
     }
 
-    #[derive(serde::Deserialize)]
-    #[serde(rename_all = "camelCase")]
-    struct WitnessOutput {
-        selected_crates: Vec<serde_json::Value>,
-    }
-
-    let output: WitnessOutput = match serde_json::from_slice(stdout) {
+    let output: PredicateOutput = match serde_json::from_slice(stdout) {
         Ok(o) => o,
         Err(e) => {
             tracing::warn!(
@@ -922,21 +898,7 @@ fn parse_witness_stdout(predicate_name: &str, stdout: &[u8]) -> Option<Vec<Witne
         }
     };
 
-    let mut crates = Vec::new();
-    for value in output.selected_crates {
-        match serde_json::from_value::<WitnessCrate>(value) {
-            Ok(wc) => crates.push(wc),
-            Err(e) => {
-                tracing::warn!(
-                    predicate = predicate_name,
-                    error = %e,
-                    "witness crate has invalid format — treating predicate as failed"
-                );
-                return None;
-            }
-        }
-    }
-    Some(crates)
+    Some(output.selected_crates)
 }
 
 #[cfg(test)]
