+++ symposium-sdk/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "symposium-sdk"
+version = "0.1.0"
+edition = "2024"
+license = "MIT OR Apache-2.0"
+description = "SDK for writing symposium plugins (hooks, custom predicates, subcommands)"
+repository = "https://github.com/symposium-dev/symposium"
+
+[dependencies]
+semver = { version = "1.0", features = ["serde"] }
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
+symposium-hook = { path = "../symposium-hook" }
